Move tags-filter unit tests into their own file.
